Add Elimination, revive, and recipe
In this update, elimination and revive is added. Recipe to craft a heart is also added. The pattern is three wither skeleton skull on the top row, soul sand, netherite block, and soul sand in the middle, and diamond block, gold block, and diamond block at the bottom.
Tbd:
Fix server resourcepack
lsadmin commands